infrastructure/tools/requirements.txt
bs4
requests
lxml
Jinja2
toml
discord.py
click